Create ability to load files into memory.
Creates what amounts to a specialized hex dump tool, but shows that the
initial needs of the program, ie, to be able to initialize a program
state and load a file into memory are set.

Test "roms" were created just as text files, one below the program limit
and one above, to test that filling the remainder with zeroes and
cutting off loading before overruning the array were successful.