Assembler partial creation.
Creation of assembler, with support for opcodes up through 0x7XNN.
These have been tested but test hasn't yet been made to automatically
check against correct results. This method of doing things is non-ideal,
of course, as it's dependent on counting character indices and any
change to my assembly languge will break this. Regex in C sounds horrid
given the reputation regexes have, and C is already bad with strings.
Rewriting the assembly mnemonics for fixed command width is "more"
appealing, but not what I am interested in, as this assembly is mainly
for testing.