Assembler created
Tested by writing an assembly "program" of every instruction one after
another, then comparing this by enabling the printf in write_op, piping
to a file, and using diff between it and a hand assembled list of
opcodes. Obviously not great but it worked and so does the assembler.

It's a very naive assembler, you can't use labels or anything; the only
nicety you get is comment lines with % (or anything up to 50 chars after
a command). You could forget the V/$ designators as long as you put
anything in that space instead. It's bad but it'll let me create some
tests slightly more easily.