Graphics working and 0x00e0
I plan to refactor the graphics portions away largely, but this should
at least remain as a known good state to return to. 0x00e0 seems to take
some time? But it's from the SDL portions. Strangely, at L98, if this
checks the state->draw_flag, the window starts with a clear surface
until 00e0 is called, and if the line is `if (1)` it shows the screen
with the test data coded into state->gfx. Why? The draw flag is
initially set to 1 anyway