0xdXYN: draw sprite opcode and font sprite data
This is the "fully featured" graphics set, further work is on
refactoring. Attempting to wrap the drawing block under an if (checking
the state->draw_flag) bricks it. it might not be a big deal? When I get
to emulating a test rom (Very soon now! Should prioritize that over
refactoring) and setting up some sort of timing, can investigate timing
requirements then.